+++ src/main/java/com/telnet/jukebox/webservice/dto/PrometDTO.java
@@ -3,24 +3,24 @@ package com.telnet.jukebox.webservice.dto;
 import java.sql.Date;
 
 public class PrometDTO {
-	private Long id;
+	private int id;
 	private Date datum;
-	private Long pesmaId;
+	private int pesmaId;
 	private String pesmaNaziv;
-	private Long cenaKolicina;
-	private Long repetition;
+	private int cenaKolicina;
+	private int repetition;
 	private String izvodjacIme;
-	private Long idKor;
+	private int idKor;
 	private String emailKor;
 
 	public PrometDTO() {
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 
@@ -40,27 +40,27 @@ public class PrometDTO {
 		this.pesmaNaziv = pesmaNaziv;
 	}
 
-	public Long getCenaKolicina() {
+	public int getCenaKolicina() {
 		return cenaKolicina;
 	}
 
-	public void setCenaKolicina(Long cenaKolicina) {
+	public void setCenaKolicina(int cenaKolicina) {
 		this.cenaKolicina = cenaKolicina;
 	}
 
-	public Long getPesmaId() {
+	public int getPesmaId() {
 		return pesmaId;
 	}
 
-	public void setPesmaId(Long pesmaId) {
+	public void setPesmaId(int pesmaId) {
 		this.pesmaId = pesmaId;
 	}
 
-	public Long getRepetition() {
+	public int getRepetition() {
 		return repetition;
 	}
 
-	public void setRepetition(Long repetition) {
+	public void setRepetition(int repetition) {
 		this.repetition = repetition;
 	}
 
@@ -80,11 +80,11 @@ public class PrometDTO {
 		this.emailKor = emailKor;
 	}
 
-	public Long getIdKor() {
+	public int getIdKor() {
 		return idKor;
 	}
 
-	public void setIdKor(Long idKor) {
+	public void setIdKor(int idKor) {
 		this.idKor = idKor;
 	}
 }
